Creates an applyPolicy function, and adds egress particle checking to it.
I'm still not particularly content with the way we match egress particles to policies, but this is the plan of record so far (and it will be easy to update later if we have to): each policy has a corresponding egress particle with the name `Egress_<PolicyName>`.

This PR adds the `applyPolicy` function, but all it does so far is check that no invalid egress particles are present in it.