Add check statements to policy translator
Adds a check statement of the form:

```
check input is allowedForEgress
  or (is allowedForEgress_redaction1 and redaction1)
  or (is allowedForEgress_redaction2 and redaction2)
```

to every readable handle of the egress particle.

Doesn't do field-level statements yet. Also no claims yet, so we can't run DFA on the resulting graphs.

I'm not heaps happy with the way I've done the unit tests here. It converts a .arcs file to a proto and loads that, and that's where I get the particle specs, recipes and policies for the test cases. I tried constructing the Kotlin data classes manually in the test file but it way very complicated and unreadable.